+++ inject.cpp
@@ -111,7 +111,7 @@ int main(int argc, char* argv[]) {
             perror("Error converting from utf8 to shift-jis");
             std::cerr << "File: " << argv[1] << std::endl;
             std::cerr << "JSON line " << json_text << std::endl;
-            for(size_t i=0;i < std::min(inbytesleft,4ul);++i)
+            for(size_t i=0;i < std::min(inbytesleft,static_cast<size_t>(4));++i)
             {
                 std::cerr << " '" << static_cast<uint8_t>(inptr[i]) << "' " << (static_cast<uint32_t>(inptr[i]) & 0x000000ff) << std::dec;
             }
